requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
